src/app/pyproject.toml
[project]
name = "library-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "advanced-alchemy>=1.8.1",
    "litestar[standard]>=2.18.0",
    "psycopg[binary,pool]>=3.3.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "sqlacodegen>=3.2.0",
]
